Move some shared functionality to convert-help.h/.cpp
---
