+++ compiler/passes/convert-help.h
@@ -111,4 +111,28 @@ struct LoopAttributeInfo {
   void insertPrimitivesBlockAtHead(UastConverter& converter, BlockStmt* body);
 };
 
+const char* convertLinkageNameAstr(const uast::Decl* node);
+Flag convertFlagForDeclLinkage(const uast::AstNode* node);
+Flag convertPragmaToFlag(uast::PragmaTag pragma);
+void attachSymbolAttributes(Context* context, const uast::Decl* node, Symbol* sym, bool containedInLibraryFile);
+void attachSymbolVisibility(const uast::Decl* node, Symbol* sym);
+UnresolvedSymExpr* reservedWordToInternalName(UniqueString name);
+Expr* reservedWordRemapForIdent(UniqueString name);
+LLVMMetadataList extractLlvmAttributesAndRejectOthers(Context* context, const uast::Loop* node);
+RetTag convertRetTag(uast::Function::ReturnIntent returnIntent);
+bool isAssignOp(UniqueString name);
+const char* createAnonymousRoutineName(const uast::Function* node);
+const char* convertFunctionNameAndAstr(const uast::Function* node);
+const char* constructUserString(const uast::Function* node);
+const char* constructUserString(const uast::FunctionSignature* node);
+IntentTag convertFormalIntent(uast::Formal::Intent intent);
+ShadowVarPrefix convertTaskVarIntent(const uast::TaskVar* node);
+const char* sanitizeVarName(const char* name, bool inTupleDecl);
+void attachSymbolStorage(const uast::Variable::Kind kind, Symbol* vs, bool setQual);
+void attachSymbolStorage(const uast::TupleDecl::IntentOrKind iok, Symbol* vs, bool setQual);
+void attachSymbolStorage(const uast::Qualifier kind, Symbol* vs, bool setQual);
+bool isEnsureDomainExprCall(Expr* expr);
+AggregateTag convertAggregateDeclTag(const uast::AggregateDecl* node);
+
+
 #endif
