+++ compiler/passes/convert-uast.cpp
@@ -68,7 +68,7 @@
 
 // If this is set then variables/formals will have their "qual" field set
 // now instead of later during resolution.
-#define ATTACH_QUALIFIED_TYPES_EARLY 0
+#define ATTACH_QUALIFIED_TYPES_EARLY false
 
 using namespace chpl;
 
@@ -226,6 +226,10 @@ struct Converter final : UastConverter {
     return shouldScopeResolve(node->id());
   }
 
+  bool isFromLibraryFile() {
+    return modStack.size() > 0 && modStack.back().isFromLibraryFile;
+  }
+
   Expr* convertExprOrNull(const uast::AstNode* node) {
     if (node == nullptr)
       return nullptr;
@@ -233,29 +237,6 @@ struct Converter final : UastConverter {
     Expr* ret = convertAST(node);
     INT_ASSERT(ret);
     return ret;
-  }
-
-  const char* convertLinkageNameAstr(const uast::Decl* node) {
-    if (auto linkageName = node->linkageName()) {
-      auto linkageStr = linkageName->toStringLiteral();
-      INT_ASSERT(linkageStr);
-      auto ret = astr(linkageStr->value());
-      return ret;
-    }
-
-    return nullptr;
-  }
-
-  Flag convertFlagForDeclLinkage(const uast::AstNode* node) {
-    if (auto decl = node->toDecl()) {
-      switch (decl->linkage()) {
-        case uast::Decl::EXTERN: return FLAG_EXTERN;
-        case uast::Decl::EXPORT: return FLAG_EXPORT;
-        default: return FLAG_UNKNOWN;
-      }
-    }
-
-    return FLAG_UNKNOWN;
   }
 
   Expr* visit(const uast::Comment* node) {
@@ -272,146 +253,8 @@ struct Converter final : UastConverter {
     return nullptr;
   }
 
-  Flag convertPragmaToFlag(chpl::uast::PragmaTag pragma) {
-    Flag ret = FLAG_UNKNOWN;
-    switch (pragma) {
-#define PRAGMA(name__, canParse__, parseStr__, desc__) \
-      case chpl::uast::PRAGMA_ ## name__ : ret = FLAG_ ## name__; break;
-#include "chpl/uast/PragmaList.h"
-#undef PRAGMA
-      default: break;
-    }
-
-    return ret;
-  }
-
-  void attachSymbolAttributes(const uast::Decl* node, Symbol* sym) {
-    if (modStack.size() > 0 && modStack.back().isFromLibraryFile) {
-      // If we are converting a symbol in a module from a .dyno
-      // file, mark the symbol as precompiled.
-      sym->addFlag(FLAG_PRECOMPILED);
-    }
-
-    const uast::AttributeGroup* attr;
-    // use the query to get the AttributeGroup or you might miss the attributes
-    // for children of multiDecl or tupleDecl
-    attr = parsing::astToAttributeGroup(context, node);
-
-    if (!attr) return;
-
-    if (attr->isDeprecated()) {
-      INT_ASSERT(!sym->hasFlag(FLAG_DEPRECATED));
-      sym->addFlag(FLAG_DEPRECATED);
-
-      auto msg = attr->deprecationMessage();
-      if (!msg.isEmpty()) {
-        sym->deprecationMsg = astr(msg);
-      }
-    } else {
-      INT_ASSERT(attr->deprecationMessage().isEmpty());
-    }
-
-    if (attr->isUnstable()) {
-      INT_ASSERT(!sym->hasFlag(FLAG_UNSTABLE));
-      sym->addFlag(FLAG_UNSTABLE);
-
-      auto msg = attr->unstableMessage();
-      if (!msg.isEmpty()) {
-        sym->unstableMsg = astr(msg);
-      }
-    } else {
-      INT_ASSERT(attr->unstableMessage().isEmpty());
-    }
-
-    if (attr->isParenfulDeprecated()) {
-      auto fnSym = toFnSymbol(sym);
-      // post-parse checks have ensured that decl is a parenless function,
-      // so the resulting symbol is a FnSymbol.
-      INT_ASSERT(fnSym);
-
-      INT_ASSERT(!sym->hasFlag(FLAG_DEPRECATED_PARENFUL));
-      sym->addFlag(FLAG_DEPRECATED_PARENFUL);
-
-      auto msg = attr->parenfulDeprecationMessage();
-      if (!msg.isEmpty()) {
-        fnSym->parenfulDeprecationMsg = astr(msg);
-      }
-    } else {
-      INT_ASSERT(attr->parenfulDeprecationMessage().isEmpty());
-    }
-
-    for (auto pragma : attr->pragmas()) {
-      Flag flag = convertPragmaToFlag(pragma);
-      if (flag != FLAG_UNKNOWN) {
-        sym->addFlag(flag);
-      }
-    }
-  }
-
-  void attachSymbolVisibility(const uast::Decl* node, Symbol* sym) {
-    if (node->visibility() == uast::Decl::PRIVATE) {
-      sym->addFlag(FLAG_PRIVATE);
-    }
-  }
-
   Expr* visit(const uast::ErroneousExpression* node) {
     return new CallExpr(PRIM_ERROR);
-  }
-
-  UnresolvedSymExpr* reservedWordToInternalName(UniqueString name) {
-    static std::unordered_map<UniqueString, const char*> map = {
-      { USTR("owned"), "_owned" },
-      { USTR("shared"), "_shared" },
-      { USTR("sync"), "_syncvar" },
-      { USTR("domain"), "_domain" },
-      { USTR("align"), "chpl_align" },
-      { USTR("by"), "chpl_by" },
-
-      // if "index" becomes an actual type, rather than magic:
-      //
-      // { USTR("index"), "_index" },
-    };
-
-    auto it = map.find(name);
-    if (it != map.end()) {
-      return new UnresolvedSymExpr(it->second);
-    }
-
-    return nullptr;
-  }
-
-  Expr* reservedWordRemapForIdent(UniqueString name) {
-    if (name == USTR("?")) {
-      return new SymExpr(gUninstantiated);
-    } else if (name == USTR("unmanaged")) {
-      return new SymExpr(dtUnmanaged->symbol);
-    } else if (name == USTR("bytes")) {
-      return new SymExpr(dtBytes->symbol);
-    } else if (name == USTR("string")) {
-      return new SymExpr(dtString->symbol);
-    } else if (name == USTR("index")) {
-      return new UnresolvedSymExpr("_index");
-    } else if (name == USTR("nil")) {
-      return new SymExpr(gNil);
-    } else if (name == USTR("noinit")) {
-      return new SymExpr(gNoInit);
-    } else if (name == USTR("locale")) {
-      return new SymExpr(dtLocale->symbol);
-    } else if (name == USTR("uint")) {
-      return new SymExpr(dtUInt[INT_SIZE_DEFAULT]->symbol);
-    } else if (name == USTR("int")) {
-      return new SymExpr(dtInt[INT_SIZE_DEFAULT]->symbol);
-    } else if (name == USTR("real")) {
-      return new SymExpr(dtReal[FLOAT_SIZE_DEFAULT]->symbol);
-    } else if (name == USTR("complex")) {
-      return new SymExpr(dtComplex[COMPLEX_SIZE_DEFAULT]->symbol);
-    } else if (name == USTR("_")) {
-      return new UnresolvedSymExpr("chpl__tuple_blank");
-    } else if (name == USTR("void")) {
-      return new SymExpr(dtVoid->symbol);
-    }
-
-    return reservedWordToInternalName(name);
   }
 
   Expr* resolvedIdentifier(const uast::Identifier* node) {
@@ -1549,23 +1392,11 @@ struct Converter final : UastConverter {
 
   /// Loops ///
 
-  LLVMMetadataList extractLlvmAttributesAndRejectOthers(const uast::Loop* node) {
-    auto loopAttributes = LoopAttributeInfo::fromExplicitLoop(context, node);
-    if (loopAttributes.assertOnGpuAttr != nullptr) {
-      CHPL_REPORT(context, InvalidGpuAttribute, node,
-                  loopAttributes.assertOnGpuAttr);
-    } else if (loopAttributes.assertEligibleAttr != nullptr) {
-      CHPL_REPORT(context, InvalidGpuAttribute, node,
-                  loopAttributes.assertEligibleAttr);
-    }
-    return std::move(loopAttributes.llvmMetadata);
-  }
-
   BlockStmt* visit(const uast::DoWhile* node) {
     Expr* condExpr = toExpr(convertAST(node->condition()));
     auto body = createBlockWithStmts(node->stmts(), node->blockStyle());
     return DoWhileStmt::build(condExpr, body,
-                              extractLlvmAttributesAndRejectOthers(node));
+                              extractLlvmAttributesAndRejectOthers(context, node));
   }
 
   BlockStmt* visit(const uast::While* node) {
@@ -1581,7 +1412,7 @@ struct Converter final : UastConverter {
     }
     auto body = createBlockWithStmts(node->stmts(), node->blockStyle());
     return WhileDoStmt::build(condExpr, body,
-                              extractLlvmAttributesAndRejectOthers(node));
+                              extractLlvmAttributesAndRejectOthers(context, node));
   }
 
   /// IndexableLoops ///
@@ -1775,7 +1606,7 @@ struct Converter final : UastConverter {
       INT_ASSERT(block);
 
       ret = ForLoop::buildForLoop(index, iteratorExpr, block, zippered,
-                                  isForExpr, extractLlvmAttributesAndRejectOthers(node));
+                                  isForExpr, extractLlvmAttributesAndRejectOthers(context, node));
     }
 
     INT_ASSERT(ret != nullptr);
@@ -2550,7 +2381,7 @@ struct Converter final : UastConverter {
 
     // TODO: Shouldn't this be all symbols?
     DefExpr* tmpDef = toDefExpr(ret->body.first());
-    attachSymbolStorage(node->intentOrKind(), tmpDef->sym);
+    attachSymbolStorage(node->intentOrKind(), tmpDef->sym, ATTACH_QUALIFIED_TYPES_EARLY);
 
     // Move the block info around like in 'buildVarDecls'.
     if (auto info = ret->blockInfoGet()) {
@@ -2693,93 +2524,6 @@ struct Converter final : UastConverter {
     return nullptr;
   }
 
-  static RetTag convertRetTag(uast::Function::ReturnIntent returnIntent) {
-    switch (returnIntent) {
-      case uast::Function::DEFAULT_RETURN_INTENT:
-      case uast::Function::OUT:
-      case uast::Function::CONST:
-        return RET_VALUE;
-      case uast::Function::CONST_REF:
-        return RET_CONST_REF;
-      case uast::Function::REF:
-        return RET_REF;
-      case uast::Function::PARAM:
-        return RET_PARAM;
-      case uast::Function::TYPE:
-        return RET_TYPE;
-    }
-
-    CHPL_UNIMPL("return intent case not handled");
-    return RET_VALUE;
-  }
-
-  static bool isAssignOp(UniqueString name) {
-    return (name == USTR("=") ||
-            name == USTR("+=") ||
-            name == USTR("-=") ||
-            name == USTR("*=") ||
-            name == USTR("/=") ||
-            name == USTR("%=") ||
-            name == USTR("**=") ||
-            name == USTR("&=") ||
-            name == USTR("|=") ||
-            name == USTR("^=") ||
-            name == USTR(">>=") ||
-            name == USTR("<<="));
-  }
-
-  static const char*
-  createAnonymousRoutineName(const uast::Function* node) {
-    std::ignore = node;
-
-    static const int maxDigits = 100;
-    static unsigned int nextId = 0;
-    static const char* prefix = "chpl_anon";
-    char buf[maxDigits];
-
-    if ((nextId + 1) == 0) INT_FATAL("Overflow for lambda ID number");
-
-    auto kind = astr(uast::Function::kindToString(node->kind()));
-
-    // Use sprintf to prevent buffer overflow if there are too many lambdas.
-    int n = snprintf(buf, (size_t) maxDigits, "%s_%s_%i", prefix, kind,
-                     nextId++);
-    if (n > (int) maxDigits) INT_FATAL("Too many lambdas.");
-
-    auto ret = astr(buf);
-    return ret;
-  }
-
-  static const char*
-  convertFunctionNameAndAstr(const uast::Function* node) {
-    auto name = node->name();
-    auto kind = node->kind();
-
-    if (node->isAnonymous()) return createAnonymousRoutineName(node);
-
-    if (name.isEmpty()) {
-      INT_ASSERT(kind == uast::Function::PROC);
-      return nullptr;
-    }
-
-    const char* ret = nullptr;
-    if (name == USTR("by")) {
-      ret = "chpl_by";
-    } else if (name == USTR("align")) {
-      ret = "chpl_align";
-    } else {
-      ret = name.c_str();
-    }
-
-    INT_ASSERT(ret);
-
-    // We have to uniquify the name now because it may be inlined (and thus
-    // stack allocated).
-    ret = astr(ret);
-
-    return ret;
-  }
-
   Expr* convertLifetimeClause(const uast::AstNode* node) {
     astlocMarker markAstLoc(node->id());
 
@@ -2822,25 +2566,6 @@ struct Converter final : UastConverter {
     return callExpr;
   }
 
-  // build up the userString as in old parser
-  // needed to match up some error outputs
-  // NOTE: parentheses may have been discarded from the original user
-  // declaration, and if so, we are not able to reconstruct them at
-  // this time
-  const char* constructUserString(const uast::Function* node) {
-    std::stringstream ss;
-    printFunctionSignature(ss, node);
-    auto ret = astr(ss.str());
-    return ret;
-  }
-
-  const char* constructUserString(const uast::FunctionSignature* node) {
-    std::stringstream ss;
-    printFunctionSignature(ss, node);
-    auto ret = astr(ss.str());
-    return ret;
-  }
-
   FnSymbol* convertFunction(const uast::Function* node) {
     // Decide if we want to resolve this function
     bool shouldScopeResolveFunction = shouldScopeResolve(node);
@@ -2869,7 +2594,7 @@ struct Converter final : UastConverter {
 
     fn->userString = constructUserString(node);
 
-    attachSymbolAttributes(node, fn);
+    attachSymbolAttributes(context, node, fn, isFromLibraryFile());
     attachSymbolVisibility(node, fn);
 
     if (node->isInline()) {
@@ -3309,7 +3034,7 @@ struct Converter final : UastConverter {
       mod->addFlag(FLAG_IMPLICIT_MODULE);
     }
 
-    attachSymbolAttributes(node, mod);
+    attachSymbolAttributes(context, node, mod, isFromLibraryFile);
 
     // Note the module is converted so we can wire up SymExprs later
     noteConvertedSym(node, mod);
@@ -3339,34 +3064,6 @@ struct Converter final : UastConverter {
   }
 
   /// VarLikeDecls ///
-
-  static IntentTag convertFormalIntent(uast::Formal::Intent intent) {
-    switch (intent) {
-      case uast::Formal::DEFAULT_INTENT:
-        return INTENT_BLANK;
-      case uast::Formal::CONST:
-        return INTENT_CONST;
-      case uast::Formal::CONST_REF:
-        return INTENT_CONST_REF;
-      case uast::Formal::REF:
-        return INTENT_REF;
-      case uast::Formal::IN:
-        return INTENT_IN;
-      case uast::Formal::CONST_IN:
-        return INTENT_CONST_IN;
-      case uast::Formal::OUT:
-        return INTENT_OUT;
-      case uast::Formal::INOUT:
-        return INTENT_INOUT;
-      case uast::Formal::PARAM:
-        return INTENT_PARAM;
-      case uast::Formal::TYPE:
-        return INTENT_TYPE;
-    }
-
-    CHPL_UNIMPL("Unhandled formal intent");
-    return INTENT_BLANK;
-  }
 
   Expr* convertTypeExpression(const uast::AstNode* node) {
     INT_ASSERT(node != nullptr);
@@ -3402,27 +3099,11 @@ struct Converter final : UastConverter {
                                 /*varargsVariable*/ nullptr);
     INT_ASSERT(ret->sym);
 
-    attachSymbolAttributes(node, ret->sym);
+    attachSymbolAttributes(context, node, ret->sym, isFromLibraryFile());
 
     // noteConvertedSym should be called when handling the enclosing Function
 
     return ret;
-  }
-
-  ShadowVarPrefix convertTaskVarIntent(const uast::TaskVar* node) {
-    astlocMarker markAstLoc(node->id());
-
-    switch (node->intent()) {
-      case uast::TaskVar::VAR: return SVP_VAR;
-      case uast::TaskVar::CONST: return SVP_CONST;
-      case uast::TaskVar::CONST_REF: return SVP_CONST_REF;
-      case uast::TaskVar::REF: return SVP_REF;
-      case uast::TaskVar::IN: return SVP_IN;
-      case uast::TaskVar::CONST_IN: return SVP_CONST_IN;
-    }
-
-    INT_FATAL("Should not reach here");
-    return SVP_VAR;
   }
 
   Expr* visit(const uast::TaskVar* node) {
@@ -3453,7 +3134,7 @@ struct Converter final : UastConverter {
                                varargsVariable);
     INT_ASSERT(ret->sym);
 
-    attachSymbolAttributes(node, ret->sym);
+    attachSymbolAttributes(context, node, ret->sym, isFromLibraryFile());
 
     return ret;
   }
@@ -3471,71 +3152,9 @@ struct Converter final : UastConverter {
 
     INT_ASSERT(ret != nullptr);
 
-    attachSymbolAttributes(node, ret);
+    attachSymbolAttributes(context, node, ret, isFromLibraryFile());
 
     return ret;
-  }
-
-  const char* sanitizeVarName(const char* name) {
-    if (inTupleDecl && name[0] == '_' && name[1] == '\0')
-      return astr("chpl__tuple_blank");
-    else
-      return astr(name);
-  }
-
-  void attachSymbolStorage(const uast::Variable::Kind kind, Symbol* vs) {
-    return attachSymbolStorage((uast::Qualifier) kind, vs);
-  }
-
-  void attachSymbolStorage(const uast::TupleDecl::IntentOrKind iok,
-                           Symbol* vs) {
-    return attachSymbolStorage((uast::Qualifier) iok, vs);
-  }
-
-  void attachSymbolStorage(const uast::Qualifier kind, Symbol* vs) {
-    auto qual = QUAL_UNKNOWN;
-
-    switch (kind) {
-      case uast::Qualifier::VAR:
-        qual = QUAL_VAL;
-        break;
-      case uast::Qualifier::CONST_VAR:
-        vs->addFlag(FLAG_CONST);
-        qual = QUAL_CONST;
-        break;
-      case uast::Qualifier::CONST_REF:
-        vs->addFlag(FLAG_CONST);
-        vs->addFlag(FLAG_REF_VAR);
-        qual = QUAL_CONST_REF;
-        break;
-      case uast::Qualifier::REF:
-        vs->addFlag(FLAG_REF_VAR);
-        qual = QUAL_REF;
-        break;
-      case uast::Qualifier::PARAM:
-        vs->addFlag(FLAG_PARAM);
-        qual = QUAL_PARAM;
-        break;
-      case uast::Qualifier::TYPE:
-        vs->addFlag(FLAG_TYPE_VARIABLE);
-        break;
-      case uast::Qualifier::INDEX:
-        vs->addFlag(FLAG_INDEX_VAR);
-        break;
-      default:
-        break;
-    }
-
-    if (ATTACH_QUALIFIED_TYPES_EARLY && qual != QUAL_UNKNOWN) {
-      vs->qual = qual;
-    }
-  }
-
-  static bool isEnsureDomainExprCall(Expr* expr) {
-    if (auto call = toCallExpr(expr)) {
-      return call->isNamed("chpl__ensureDomainExpr");
-    }
-    return false;
   }
 
   CallExpr* convertArrayType(const uast::BracketLoop* node) {
@@ -3710,7 +3329,7 @@ struct Converter final : UastConverter {
                     (multiState != nullptr && multiState->localeTemp != nullptr);
     auto block = (isRemote || multiState) ? new BlockStmt(BLOCK_SCOPELESS) : nullptr;
 
-    auto varSym = new VarSymbol(sanitizeVarName(node->name().c_str()));
+    auto varSym = new VarSymbol(sanitizeVarName(node->name().c_str(), inTupleDecl));
     const bool isTypeVar = node->kind() == uast::Variable::TYPE;
 
     if (fIdBasedMunging && node->linkage() == uast::Decl::DEFAULT_LINKAGE) {
@@ -3763,9 +3382,9 @@ struct Converter final : UastConverter {
     }
 
     // Adjust the variable according to its kind, e.g. 'const'/'type'.
-    attachSymbolStorage(symbolKind, varSym);
+    attachSymbolStorage(symbolKind, varSym, ATTACH_QUALIFIED_TYPES_EARLY);
 
-    attachSymbolAttributes(node, varSym);
+    attachSymbolAttributes(context, node, varSym, isFromLibraryFile());
 
     attachSymbolVisibility(node, varSym);
 
@@ -3968,7 +3587,7 @@ struct Converter final : UastConverter {
     const char* name = astr(node->name());
     Expr* initExpr = convertExprOrNull(node->initExpression());
     auto ret = new DefExpr(new EnumSymbol(name), initExpr);
-    attachSymbolAttributes(node, ret->sym);
+    attachSymbolAttributes(context, node, ret->sym, isFromLibraryFile());
 
     // Note the enum element is converted so we can wire up SymExprs later
     noteConvertedSym(node, ret->sym);
@@ -3997,7 +3616,7 @@ struct Converter final : UastConverter {
 
     auto enumTypeSym = new TypeSymbol(node->name().c_str(), enumType);
 
-    attachSymbolAttributes(node, enumTypeSym);
+    attachSymbolAttributes(context, node, enumTypeSym, isFromLibraryFile());
     attachSymbolVisibility(node, enumTypeSym);
 
     enumType->symbol = enumTypeSym;
@@ -4014,17 +3633,6 @@ struct Converter final : UastConverter {
   }
 
   /// AggregateDecls
-
-  AggregateTag convertAggregateDeclTag(const uast::AggregateDecl* node) {
-    switch (node->tag()) {
-      case uast::asttags::Class: return AGGREGATE_CLASS;
-      case uast::asttags::Record: return AGGREGATE_RECORD;
-      case uast::asttags::Union: return AGGREGATE_UNION;
-      default: break;
-    }
-    INT_FATAL("Should not reach here!");
-    return AGGREGATE_CLASS;
-  }
 
   template <typename Iterable>
   void convertInheritsExprs(const Iterable& iterable,
@@ -4103,7 +3711,7 @@ struct Converter final : UastConverter {
                                  topLevelModTag);
     INT_ASSERT(ret->sym);
 
-    attachSymbolAttributes(node, ret->sym);
+    attachSymbolAttributes(context, node, ret->sym, isFromLibraryFile());
     attachSymbolVisibility(node, ret->sym);
     if (inheritMarkedGeneric) {
       ret->sym->addFlag(FLAG_SUPERCLASS_MARKED_GENERIC);
